UserAuthenticationProvider added to read and verify credentials from openmrs users table
