+++ api/src/main/java/org/openmrs/module/oauth2/api/impl/ClientAuthenticationService.java
@@ -1,0 +1,7 @@
+package org.openmrs.module.oauth2.api.impl;
+
+/**
+ * Created by OPSKMC on 8/20/15.
+ */
+public class ClientAuthenticationService {
+}
