--- api/src/main/java/org/openmrs/module/oauth2/api/impl/UserAuthenticationProvider.java
@@ -1,26 +1,0 @@
-package org.openmrs.module.oauth2.api.impl;
-
-import org.openmrs.api.UserService;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-
-/**
- * Created by OPSKMC on 8/18/15.
- */
-public class UserAuthenticationProvider implements AuthenticationProvider {
-
-    @Override
-    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-        String username = authentication.getName();
-        String password = authentication.getCredentials().toString();
-        // Context.
-        return null;
-    }
-
-    @Override
-    public boolean supports(Class<?> authentication) {
-        return (UsernamePasswordAuthenticationToken.class.isAssignableFrom(authentication));
-    }
-}
