+++ omod/src/main/resources/webModuleApplicationContext.xml
@@ -37,13 +37,13 @@
 
     <!--
 
-        <security:http pattern="/module/**" use-expressions="true"
-                       authentication-manager-ref="authenticationManager">
-                <security:intercept-url pattern="/module/oauth2/**" access="hasRole('ROLE_USER')"/>
-                <security:intercept-url pattern="/module/fhir/**" access="hasRole('ROLE_USER')"/>
-                <security:intercept-url pattern="/module/basicmodule/**" access="hasRole('ROLE_USER')"/>
-                <security:form-login/>
-            </security:http>
+            <security:http pattern="/module/**" use-expressions="true"
+                           authentication-manager-ref="authenticationManager">
+                    <security:intercept-url pattern="/module/oauth2/**" access="hasRole('ROLE_USER')"/>
+                    <security:intercept-url pattern="/module/fhir/**" access="hasRole('ROLE_USER')"/>
+                    <security:intercept-url pattern="/module/basicmodule/**" access="hasRole('ROLE_USER')"/>
+                    <security:form-login/>
+                </security:http>
 
     -->
 
@@ -54,13 +54,13 @@
     <!--via Authorization Code/ Implicit Grant Type-->
     <!--===================================================================-->
     <security:authentication-manager alias="authenticationManager">
-        <security:authentication-provider>
-            <security:user-service id="userDetailsService">
-                <security:user name="bob" authorities="ROLE_USER" password="bob"/>
-            </security:user-service>
+        <security:authentication-provider ref="userAuthenticationProvider">
+            <!--       <security:user-service id="userDetailsService">
+                       <security:user name="bob" authorities="ROLE_USER" password="bob"/>
+                   </security:user-service>-->
         </security:authentication-provider>
     </security:authentication-manager>
-    <!--<bean id="userAuthenticationProvider" class=""/>-->
+    <bean id="userAuthenticationProvider" class="org.openmrs.module.oauth2.api.impl.UserAuthenticationService"/>
     <!--============================-->
     <!--OAuth2 Entry Points. What is Realm Name??
         @see http://stackoverflow.com/questions/10892336/realm-name-in-tomcat-web-xml
